States class, OFF class & init
Introduced the abstract base class "States" from which the ambient, turbine and OP classes inherit. The class offers basic set, get, iterate and init functionality, which avoids redundancy for the derived states and a joined method names.

OFF class is the simulation object which initialises the simulation and eventually will run it.

Introduced an init functionality which gives all states a starting value and orientates the OPs downstream.

Also added a general toolbox object "OFFTools" which is meant to host practical functions. Don't know how I feel about it yet