OFF Tools & docstrings
Changed OFFTools to a function collection, and changed a lot of docstrings to the numpy docstring convention.

Also added a states object attribute which requires a list of the names for the states.

Adapted the ambient class to have a function for wind speed, absolute, u & v component - makes conversion easier