run_id fix
Path retrieval did not work on my system as the key "OFF_PATH" was not part of os.environ - might be due to the IDE used? How does a stand alone execution do it?
Replaced it with a try/except block with entry that does work.

Maybe in long term better to retrieve file path and reduce to program path.