FLORIS yaml & cleanup
Added FLORIS .yaml files incl. attribution to the NREL git. This way, the user does not need to reference a local path to a .yaml file but the code can reference it directly.

Linked the new settings to be used in the FLORIS wake model Chris built. Also cleaned up a bit & added comments