run file, inputs, turbine
Added a run .yaml file as example, which is the copy-pasted version of the used data base entries.

modified main to read the new run file and to convert it into the settings files.

The DTU10MW turbine object was changed to be a generic HAWT turbine. The input file can later define what kind of model is added (VAWT, HAWT floating, VAWT floating, airborne,...)