Fixed a bug where the measurement turbine was a reference of a "true" turbine obj
The measurement turbine was a reference of the turbine at index 0, so all changes applied to the measurement turbine were also applied to turbine 0. This had no effect on the plot, but might have very well changed the further simulation results.

The stored plot is also fixed now and does not appear blank